Create Hacker Rank For loop with Switch case.cpp
Hacker Rank C++ For Loop Question Solution: You will be given two positive integers, a and b, separated by a newline. Loop from a to b, and check if the num its smaller than 9 print the English representation of it in lowercase, or if it is greater than 9, print if its even or odd.